requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
transformers
huggingface_hub
google-api-python-client 